requirements.txt
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7
streamlit==1.29.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.0
nest_asyncio==1.6.0
tensorflow==2.12.0
